third_party/WebKit/LayoutTests/inspector/elements/styles-1/edit-media-text-expected.txt

Tests that editing media text updates element styles.

Text
=== Before media text modification ===
[expanded] 
element.style { ()
    color: red;

[expanded] 
@media screen and (max-device-width: 100000px)
#inspected { (edit-media-text.html:8 -> edit-media-text.html:8:17)
/-- overloaded --/     color: blue;

[expanded] 
@media screen and (max-device-width: 100000px)
#inspected { (edit-media-text.html:5 -> edit-media-text.html:5:17)
/-- overloaded --/     color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;

=== After valid media text modification ===
[expanded] 
element.style { ()
    color: red;

[expanded] 
@media screen and (max-device-width: 99999px)
#inspected { (edit-media-text.html:8 -> edit-media-text.html:8:17)
/-- overloaded --/     color: blue;

[expanded] 
@media screen and (max-device-width: 99999px)
#inspected { (edit-media-text.html:5 -> edit-media-text.html:5:17)
/-- overloaded --/     color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;

=== After invalid media text modification ===
[expanded] 
element.style { ()
    color: red;

[expanded] 
#inspected { (edit-media-text.html:8 -> edit-media-text.html:8:17)
/-- overloaded --/     color: blue;

[expanded] 
#inspected { (edit-media-text.html:5 -> edit-media-text.html:5:17)
/-- overloaded --/     color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;
